# internal/entities/user.go
package entities

// User model
type User struct {
	Model
	Username  string `json:"username"`
	Email     string `json:"email"`
	Password  string `json:"password"`
	Firstname string `json:"firstname"`
	Lastname  string `json:"lastname"`
} //@name User